Homoiconicity
Updates data structures such that a parsed SLOP
program is represented using native SLOP runtime
data types such as `list`, `symbol`, `number`,
etc.

Along that same line, simplifies these data
structures such that we're as close to `1:1` with
`SLOP-native:Js-native` types. A list is a JS
array, a dict is a JS object, etc. The one place
we had to diverge slightly is with symbols and
keywords.

SLOP types can be tested via `SlopPred`,
constructed via `SlopVal`, and maninpulated via
`Slop<Type>` APIs e.g. `SlopList`, all opaquely.
In retrospect, the utility of the `Slop<Type>`
APIs is questionable, since most of their
contracts is that they're 1-1 with JS types. This
way at least, we have some room for divergence.